requirements.txt
flask
transformers
torch
google-generativeai
python-dotenv
pandas
beautifulsoup4
requests
